requirements.txt
openai==0.27.0
urllib3==1.26.14
Pillow==9.2.0
requests==2.28.2